requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
frozenlist==1.3.3
hiredis==2.1.0
idna==3.4
magic-filter==1.0.9
more-itertools==8.14.0
multidict==6.0.3
pptree==3.1
pydantic==1.10.2
python-ulid==1.1.0
pytz==2022.6
redis==4.4.0
redis-om==0.1.1
types-redis==4.3.21.6
typing_extensions==4.4.0
yarl==1.8.2
